requirements/default.txt
django-service-templates==1.0.0.0a2
djangorestframework==3.3.1
-r graphql.txt